added put and patch mapping for full and partial updates , tip: always check if entity exists first

+++ src/main/java/com/example/ecommerce/Service/UserService.java
@@ -6,8 +6,11 @@ import com.example.ecommerce.Repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+import org.springframework.util.ReflectionUtils;
 
+import java.lang.reflect.Field;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
@@ -19,7 +22,9 @@ public class UserService {
 
 
     public Boolean deleteById(Long id) {
-         userRepository.deleteById(id);
+        boolean isExist = userRepository.existsById(id);
+        if(!isExist) return false;
+        userRepository.deleteById(id);
          return true;
     }
 
@@ -40,5 +45,31 @@ public class UserService {
         User user = modelMapper.map(dto , User.class);
         User saved = userRepository.save(user);
         return modelMapper.map(saved , UserDto.class);
+    }
+
+    public UserDto updateUserById(Long id, UserDto userDto) throws Exception {
+        boolean isExist = userRepository.existsById(id);
+        if(!isExist) throw new Exception("User does not exist");
+
+        User user = modelMapper.map(userDto , User.class);
+        return modelMapper.map(userRepository.save(user) , UserDto.class);
+    }
+
+
+    public UserDto partiallyUpdateUserById(Long id, Map<String, Object> map) throws Exception {
+
+        boolean isExist = userRepository.existsById(id);
+        if(!isExist) throw new Exception("employee does not exist");
+
+        User user = userRepository.findById(id).orElseThrow();
+
+        map.forEach((key, value) -> {
+            Field field = ReflectionUtils.findField(User.class , key);
+            field.setAccessible(true);
+            ReflectionUtils.setField(field,user,value);
+        });
+        User savedUser = userRepository.save(user);
+        return modelMapper.map(savedUser,UserDto.class);
+
     }
 }
